src/scripts/requirements.txt
Flask_Cors==3.0.10
mysql_connector_repackaged==0.3.1
python-dotenv==0.19.2
